Crate rusoto_emr

Amazon EMR is a web service that makes it easy to process large amounts of data efficiently. Amazon EMR uses Hadoop processing combined with several AWS products to do tasks such as web indexing, data mining, log file analysis, machine learning, scientific simulation, and data warehousing.

If you're using the service, you're probably looking for EmrClient and Emr.

Structs

AddInstanceFleetInput

AddInstanceFleetOutput

AddInstanceGroupsInput
Input to an AddInstanceGroups call.

AddInstanceGroupsOutput
Output from an AddInstanceGroups call.

AddJobFlowStepsInput
The input argument to the AddJobFlowSteps operation.

AddJobFlowStepsOutput
The output for the AddJobFlowSteps operation.

AddTagsInput
This input identifies a cluster and a list of tags to attach.

AddTagsOutput
This output indicates the result of adding tags to a resource.

Application
With Amazon EMR release version 4.0 and later, the only accepted parameter is the application name. To pass arguments to applications, you use configuration classifications specified using configuration JSON objects. For more information, see Configuring Applications.

With earlier Amazon EMR releases, the application is any Amazon or third-party software that you can add to the cluster. This structure contains a list of strings that indicates the software to use with the cluster and accepts a user argument list. Amazon EMR accepts and forwards the argument list to the corresponding installation script as bootstrap action argument.

AutoScalingPolicy
An automatic scaling policy for a core instance group or task instance group in an Amazon EMR cluster. An automatic scaling policy defines how an instance group dynamically adds and terminates EC2 instances in response to the value of a CloudWatch metric. See PutAutoScalingPolicy.

AutoScalingPolicyDescription
An automatic scaling policy for a core instance group or task instance group in an Amazon EMR cluster. The automatic scaling policy defines how an instance group dynamically adds and terminates EC2 instances in response to the value of a CloudWatch metric. See PutAutoScalingPolicy.

AutoScalingPolicyStateChangeReason
The reason for an AutoScalingPolicyStatus change.

AutoScalingPolicyStatus
The status of an automatic scaling policy.

BootstrapActionConfig
Configuration of a bootstrap action.

BootstrapActionDetail
Reports the configuration of a bootstrap action in a cluster (job flow).

CancelStepsInfo
Specification of the status of a CancelSteps request. Available only in Amazon EMR version 4.8.0 and later, excluding version 5.0.0.

CancelStepsInput
The input argument to the CancelSteps operation.

CancelStepsOutput
The output for the CancelSteps operation.

CloudWatchAlarmDefinition
The definition of a CloudWatch metric alarm, which determines when an automatic scaling activity is triggered. When the defined alarm conditions are satisfied, scaling activity begins.

Cluster
The detailed description of the cluster.

ClusterStateChangeReason
The reason that the cluster changed to its current state.

ClusterStatus
The detailed status of the cluster.

ClusterSummary
The summary description of the cluster.

ClusterTimeline
Represents the timeline of the cluster's lifecycle.

Command
An entity describing an executable that runs on a cluster.

Configuration

Amazon EMR releases 4.x or later.

An optional configuration specification to be used when provisioning cluster instances, which can include configurations for applications and software bundled with Amazon EMR. A configuration consists of a classification, properties, and optional nested configurations. A classification refers to an application-specific configuration file. Properties are the settings you want to change in that file. For more information, see Configuring Applications.

CreateSecurityConfigurationInput

CreateSecurityConfigurationOutput

DeleteSecurityConfigurationInput

DeleteSecurityConfigurationOutput

DescribeClusterInput
This input determines which cluster to describe.

DescribeClusterOutput
This output contains the description of the cluster.

DescribeJobFlowsInput
The input for the DescribeJobFlows operation.

DescribeJobFlowsOutput
The output for the DescribeJobFlows operation.

DescribeSecurityConfigurationInput

DescribeSecurityConfigurationOutput

DescribeStepInput
This input determines which step to describe.

DescribeStepOutput
This output contains the description of the cluster step.

EbsBlockDevice
Configuration of requested EBS block device associated with the instance group.

EbsBlockDeviceConfig
Configuration of requested EBS block device associated with the instance group with count of volumes that will be associated to every instance.

EbsConfiguration
The Amazon EBS configuration of a cluster instance.

EbsVolume
EBS block device that's attached to an EC2 instance.

Ec2InstanceAttributes
Provides information about the EC2 instances in a cluster grouped by category. For example, key name, subnet ID, IAM instance profile, and so on.

EmrClient
A client for the Amazon EMR API.

FailureDetails
The details of the step failure. The service attempts to detect the root cause for many common failures.

HadoopJarStepConfig
A job flow step consisting of a JAR file whose main function will be executed. The main function submits a job for Hadoop to execute and waits for the job to finish or fail.

HadoopStepConfig
A cluster step consisting of a JAR file whose main function will be executed. The main function submits a job for Hadoop to execute and waits for the job to finish or fail.

Instance
Represents an EC2 instance provisioned as part of cluster.

InstanceFleet

Describes an instance fleet, which is a group of EC2 instances that host a particular node type (master, core, or task) in an Amazon EMR cluster. Instance fleets can consist of a mix of instance types and On-Demand and Spot instances, which are provisioned to meet a defined target capacity.

The instance fleet configuration is available only in Amazon EMR versions 4.8.0 and later, excluding 5.0.x versions.

InstanceFleetConfig

The configuration that defines an instance fleet.

The instance fleet configuration is available only in Amazon EMR versions 4.8.0 and later, excluding 5.0.x versions.

InstanceFleetModifyConfig

Configuration parameters for an instance fleet modification request.

The instance fleet configuration is available only in Amazon EMR versions 4.8.0 and later, excluding 5.0.x versions.

InstanceFleetProvisioningSpecifications

The launch specification for Spot instances in the fleet, which determines the defined duration and provisioning timeout behavior.

The instance fleet configuration is available only in Amazon EMR versions 4.8.0 and later, excluding 5.0.x versions.

InstanceFleetStateChangeReason

Provides status change reason details for the instance fleet.

The instance fleet configuration is available only in Amazon EMR versions 4.8.0 and later, excluding 5.0.x versions.

InstanceFleetStatus

The status of the instance fleet.

The instance fleet configuration is available only in Amazon EMR versions 4.8.0 and later, excluding 5.0.x versions.

InstanceFleetTimeline

Provides historical timestamps for the instance fleet, including the time of creation, the time it became ready to run jobs, and the time of termination.

The instance fleet configuration is available only in Amazon EMR versions 4.8.0 and later, excluding 5.0.x versions.

InstanceGroup
This entity represents an instance group, which is a group of instances that have common purpose. For example, CORE instance group is used for HDFS.

InstanceGroupConfig
Configuration defining a new instance group.

InstanceGroupDetail
Detailed information about an instance group.

InstanceGroupModifyConfig
Modify the size or configurations of an instance group.

InstanceGroupStateChangeReason
The status change reason details for the instance group.

InstanceGroupStatus
The details of the instance group status.

InstanceGroupTimeline
The timeline of the instance group lifecycle.

InstanceResizePolicy
Custom policy for requesting termination protection or termination of specific instances when shrinking an instance group.

InstanceStateChangeReason
The details of the status change reason for the instance.

InstanceStatus
The instance status details.

InstanceTimeline
The timeline of the instance lifecycle.

InstanceTypeConfig

An instance type configuration for each instance type in an instance fleet, which determines the EC2 instances Amazon EMR attempts to provision to fulfill On-Demand and Spot target capacities. There can be a maximum of 5 instance type configurations in a fleet.

The instance fleet configuration is available only in Amazon EMR versions 4.8.0 and later, excluding 5.0.x versions.

InstanceTypeSpecification

The configuration specification for each instance type in an instance fleet.

The instance fleet configuration is available only in Amazon EMR versions 4.8.0 and later, excluding 5.0.x versions.

JobFlowDetail
A description of a cluster (job flow).

JobFlowExecutionStatusDetail
Describes the status of the cluster (job flow).

JobFlowInstancesConfig
A description of the Amazon EC2 instance on which the cluster (job flow) runs. A valid JobFlowInstancesConfig must contain either InstanceGroups or InstanceFleets, which is the recommended configuration. They cannot be used together. You may also have MasterInstanceType, SlaveInstanceType, and InstanceCount (all three must be present), but we don't recommend this configuration.

JobFlowInstancesDetail
Specify the type of Amazon EC2 instances that the cluster (job flow) runs on.

KerberosAttributes
Attributes for Kerberos configuration when Kerberos authentication is enabled using a security configuration. For more information see Use Kerberos Authentication in the EMR Management Guide.

KeyValue
A key value pair.

ListBootstrapActionsInput
This input determines which bootstrap actions to retrieve.

ListBootstrapActionsOutput
This output contains the bootstrap actions detail.

ListClustersInput
This input determines how the ListClusters action filters the list of clusters that it returns.

ListClustersOutput
This contains a ClusterSummaryList with the cluster details; for example, the cluster IDs, names, and status.

ListInstanceFleetsInput

ListInstanceFleetsOutput

ListInstanceGroupsInput
This input determines which instance groups to retrieve.

ListInstanceGroupsOutput
This input determines which instance groups to retrieve.

ListInstancesInput
This input determines which instances to list.

ListInstancesOutput
This output contains the list of instances.

ListSecurityConfigurationsInput

ListSecurityConfigurationsOutput

ListStepsInput
This input determines which steps to list.

ListStepsOutput
This output contains the list of steps returned in reverse order. This means that the last step is the first element in the list.

MetricDimension
A CloudWatch dimension, which is specified using a Key (known as a Name in CloudWatch), Value pair. By default, Amazon EMR uses one dimension whose Key is JobFlowID and Value is a variable representing the cluster ID, which is ${emr.clusterId}. This enables the rule to bootstrap when the cluster ID becomes available.

ModifyInstanceFleetInput

ModifyInstanceGroupsInput
Change the size of some instance groups.

PlacementType
The Amazon EC2 Availability Zone configuration of the cluster (job flow).

PutAutoScalingPolicyInput

PutAutoScalingPolicyOutput

RemoveAutoScalingPolicyInput

RemoveAutoScalingPolicyOutput

RemoveTagsInput
This input identifies a cluster and a list of tags to remove.

RemoveTagsOutput
This output indicates the result of removing tags from a resource.

RunJobFlowInput
Input to the RunJobFlow operation.

RunJobFlowOutput
The result of the RunJobFlow operation.

ScalingAction
The type of adjustment the automatic scaling activity makes when triggered, and the periodicity of the adjustment.

ScalingConstraints
The upper and lower EC2 instance limits for an automatic scaling policy. Automatic scaling activities triggered by automatic scaling rules will not cause an instance group to grow above or below these limits.

ScalingRule
A scale-in or scale-out rule that defines scaling activity, including the CloudWatch metric alarm that triggers activity, how EC2 instances are added or removed, and the periodicity of adjustments. The automatic scaling policy for an instance group can comprise one or more automatic scaling rules.

ScalingTrigger
The conditions that trigger an automatic scaling activity.

ScriptBootstrapActionConfig
Configuration of the script to run during a bootstrap action.

SecurityConfigurationSummary
The creation date and time, and name, of a security configuration.

SetTerminationProtectionInput
The input argument to the TerminationProtection operation.

SetVisibleToAllUsersInput
The input to the SetVisibleToAllUsers action.

ShrinkPolicy
Policy for customizing shrink operations. Allows configuration of decommissioning timeout and targeted instance shrinking.

SimpleScalingPolicyConfiguration
An automatic scaling configuration, which describes how the policy adds or removes instances, the cooldown period, and the number of EC2 instances that will be added each time the CloudWatch metric alarm condition is satisfied.

SpotProvisioningSpecification

The launch specification for Spot instances in the instance fleet, which determines the defined duration and provisioning timeout behavior.

The instance fleet configuration is available only in Amazon EMR versions 4.8.0 and later, excluding 5.0.x versions.

Step
This represents a step in a cluster.

StepConfig
Specification of a cluster (job flow) step.

StepDetail
Combines the execution state and configuration of a step.

StepExecutionStatusDetail
The execution state of a step.

StepStateChangeReason
The details of the step state change reason.

StepStatus
The execution status details of the cluster step.

StepSummary
The summary of the cluster step.

StepTimeline
The timeline of the cluster step lifecycle.

SupportedProductConfig
The list of supported product configurations which allow user-supplied arguments. EMR accepts these arguments and forwards them to the corresponding installation script as bootstrap action arguments.

Tag
A key/value pair containing user-defined metadata that you can associate with an Amazon EMR resource. Tags make it easier to associate clusters in various ways, such as grouping clusters to track your Amazon EMR resource allocation costs. For more information, see Tag Clusters.

TerminateJobFlowsInput
Input to the TerminateJobFlows operation.

VolumeSpecification
EBS volume specifications such as volume type, IOPS, and size (GiB) that will be requested for the EBS volume attached to an EC2 instance in the cluster.

Enums

AddInstanceFleetError
Errors returned by AddInstanceFleet

AddInstanceGroupsError
Errors returned by AddInstanceGroups

AddJobFlowStepsError
Errors returned by AddJobFlowSteps

AddTagsError
Errors returned by AddTags

CancelStepsError
Errors returned by CancelSteps

CreateSecurityConfigurationError
Errors returned by CreateSecurityConfiguration

DeleteSecurityConfigurationError
Errors returned by DeleteSecurityConfiguration

DescribeClusterError
Errors returned by DescribeCluster

DescribeJobFlowsError
Errors returned by DescribeJobFlows

DescribeSecurityConfigurationError
Errors returned by DescribeSecurityConfiguration

DescribeStepError
Errors returned by DescribeStep

ListBootstrapActionsError
Errors returned by ListBootstrapActions

ListClustersError
Errors returned by ListClusters

ListInstanceFleetsError
Errors returned by ListInstanceFleets

ListInstanceGroupsError
Errors returned by ListInstanceGroups

ListInstancesError
Errors returned by ListInstances

ListSecurityConfigurationsError
Errors returned by ListSecurityConfigurations

ListStepsError
Errors returned by ListSteps

ModifyInstanceFleetError
Errors returned by ModifyInstanceFleet

ModifyInstanceGroupsError
Errors returned by ModifyInstanceGroups

PutAutoScalingPolicyError
Errors returned by PutAutoScalingPolicy

RemoveAutoScalingPolicyError
Errors returned by RemoveAutoScalingPolicy

RemoveTagsError
Errors returned by RemoveTags

RunJobFlowError
Errors returned by RunJobFlow

SetTerminationProtectionError
Errors returned by SetTerminationProtection

SetVisibleToAllUsersError
Errors returned by SetVisibleToAllUsers

TerminateJobFlowsError
Errors returned by TerminateJobFlows

Traits

Emr
Trait representing the capabilities of the Amazon EMR API. Amazon EMR clients implement this trait.