Struct rusoto_emr::CloudWatchAlarmDefinition

pub struct CloudWatchAlarmDefinition {
    pub comparison_operator: String,
    pub dimensions: Option<Vec<MetricDimension>>,
    pub evaluation_periods: Option<i64>,
    pub metric_name: String,
    pub namespace: Option<String>,
    pub period: i64,
    pub statistic: Option<String>,
    pub threshold: f64,
    pub unit: Option<String>,
}

The definition of a CloudWatch metric alarm, which determines when an automatic scaling activity is triggered. When the defined alarm conditions are satisfied, scaling activity begins.

Fields
comparison_operator: String

Determines how the metric specified by MetricName is compared to the value specified by Threshold.

dimensions: Option<Vec<MetricDimension>>

A CloudWatch metric dimension.

evaluation_periods: Option<i64>

The number of periods, expressed in seconds using Period, during which the alarm condition must exist before the alarm triggers automatic scaling activity. The default value is 1.

metric_name: String

The name of the CloudWatch metric that is watched to determine an alarm condition.

namespace: Option<String>

The namespace for the CloudWatch metric. The default is AWS/ElasticMapReduce.

period: i64

The period, in seconds, over which the statistic is applied. EMR CloudWatch metrics are emitted every five minutes (300 seconds), so if an EMR CloudWatch metric is specified, specify 300.

statistic: Option<String>

The statistic to apply to the metric associated with the alarm. The default is AVERAGE.

threshold: f64

The value against which the specified statistic is compared.

unit: Option<String>

The unit of measure associated with the CloudWatch metric being watched. The value specified for Unit must correspond to the units specified in the CloudWatch metric.