Struct rusoto_emr::InstanceFleetConfig

pub struct InstanceFleetConfig {
    pub instance_fleet_type: String,
    pub instance_type_configs: Option<Vec<InstanceTypeConfig>>,
    pub launch_specifications: Option<InstanceFleetProvisioningSpecifications>,
    pub name: Option<String>,
    pub target_on_demand_capacity: Option<i64>,
    pub target_spot_capacity: Option<i64>,
}

The configuration that defines an instance fleet.

The instance fleet configuration is available only in Amazon EMR versions 4.8.0 and later, excluding 5.0.x versions.

Fields
instance_fleet_type: String

The node type that the instance fleet hosts. Valid values are MASTER,CORE,and TASK.

instance_type_configs: Option<Vec<InstanceTypeConfig>>

The instance type configurations that define the EC2 instances in the instance fleet.

launch_specifications: Option<InstanceFleetProvisioningSpecifications>

The launch specification for the instance fleet.

name: Option<String>

The friendly name of the instance fleet.

target_on_demand_capacity: Option<i64>

The target capacity of On-Demand units for the instance fleet, which determines how many On-Demand instances to provision. When the instance fleet launches, Amazon EMR tries to provision On-Demand instances as specified by InstanceTypeConfig. Each instance configuration has a specified WeightedCapacity. When an On-Demand instance is provisioned, the WeightedCapacity units count toward the target capacity. Amazon EMR provisions instances until the target capacity is totally fulfilled, even if this results in an overage. For example, if there are 2 units remaining to fulfill capacity, and Amazon EMR can only provision an instance with a WeightedCapacity of 5 units, the instance is provisioned, and the target capacity is exceeded by 3 units.

If not specified or set to 0, only Spot instances are provisioned for the instance fleet using TargetSpotCapacity. At least one of TargetSpotCapacity and TargetOnDemandCapacity should be greater than 0. For a master instance fleet, only one of TargetSpotCapacity and TargetOnDemandCapacity can be specified, and its value must be 1.

target_spot_capacity: Option<i64>

The target capacity of Spot units for the instance fleet, which determines how many Spot instances to provision. When the instance fleet launches, Amazon EMR tries to provision Spot instances as specified by InstanceTypeConfig. Each instance configuration has a specified WeightedCapacity. When a Spot instance is provisioned, the WeightedCapacity units count toward the target capacity. Amazon EMR provisions instances until the target capacity is totally fulfilled, even if this results in an overage. For example, if there are 2 units remaining to fulfill capacity, and Amazon EMR can only provision an instance with a WeightedCapacity of 5 units, the instance is provisioned, and the target capacity is exceeded by 3 units.

If not specified or set to 0, only On-Demand instances are provisioned for the instance fleet. At least one of TargetSpotCapacity and TargetOnDemandCapacity should be greater than 0. For a master instance fleet, only one of TargetSpotCapacity and TargetOnDemandCapacity can be specified, and its value must be 1.