Struct rusoto_emr::InstanceGroup

pub struct InstanceGroup {
    pub auto_scaling_policy: Option<AutoScalingPolicyDescription>,
    pub bid_price: Option<String>,
    pub configurations: Option<Vec<Configuration>>,
    pub ebs_block_devices: Option<Vec<EbsBlockDevice>>,
    pub ebs_optimized: Option<bool>,
    pub id: Option<String>,
    pub instance_group_type: Option<String>,
    pub instance_type: Option<String>,
    pub market: Option<String>,
    pub name: Option<String>,
    pub requested_instance_count: Option<i64>,
    pub running_instance_count: Option<i64>,
    pub shrink_policy: Option<ShrinkPolicy>,
    pub status: Option<InstanceGroupStatus>,
}

This entity represents an instance group, which is a group of instances that have common purpose. For example, CORE instance group is used for HDFS.

Fields
auto_scaling_policy: Option<AutoScalingPolicyDescription>

An automatic scaling policy for a core instance group or task instance group in an Amazon EMR cluster. The automatic scaling policy defines how an instance group dynamically adds and terminates EC2 instances in response to the value of a CloudWatch metric. See PutAutoScalingPolicy.

bid_price: Option<String>

The bid price for each EC2 instance in the instance group when launching nodes as Spot Instances, expressed in USD.

configurations: Option<Vec<Configuration>>

Amazon EMR releases 4.x or later.

The list of configurations supplied for an EMR cluster instance group. You can specify a separate configuration for each instance group (master, core, and task).

ebs_block_devices: Option<Vec<EbsBlockDevice>>

The EBS block devices that are mapped to this instance group.

ebs_optimized: Option<bool>

If the instance group is EBS-optimized. An Amazon EBS-optimized instance uses an optimized configuration stack and provides additional, dedicated capacity for Amazon EBS I/O.

id: Option<String>

The identifier of the instance group.

instance_group_type: Option<String>

The type of the instance group. Valid values are MASTER, CORE or TASK.

instance_type: Option<String>

The EC2 instance type for all instances in the instance group.

market: Option<String>

The marketplace to provision instances for this group. Valid values are ON_DEMAND or SPOT.

name: Option<String>

The name of the instance group.

requested_instance_count: Option<i64>

The target number of instances for the instance group.

running_instance_count: Option<i64>

The number of instances currently running in this instance group.

shrink_policy: Option<ShrinkPolicy>

Policy for customizing shrink operations.

status: Option<InstanceGroupStatus>

The current status of the instance group.