Struct rusoto_emr::InstanceGroupConfig

pub struct InstanceGroupConfig {
    pub auto_scaling_policy: Option<AutoScalingPolicy>,
    pub bid_price: Option<String>,
    pub configurations: Option<Vec<Configuration>>,
    pub ebs_configuration: Option<EbsConfiguration>,
    pub instance_count: i64,
    pub instance_role: String,
    pub instance_type: String,
    pub market: Option<String>,
    pub name: Option<String>,
}

Configuration defining a new instance group.

Fields
auto_scaling_policy: Option<AutoScalingPolicy>

An automatic scaling policy for a core instance group or task instance group in an Amazon EMR cluster. The automatic scaling policy defines how an instance group dynamically adds and terminates EC2 instances in response to the value of a CloudWatch metric. See PutAutoScalingPolicy.

bid_price: Option<String>

Bid price for each EC2 instance in the instance group when launching nodes as Spot Instances, expressed in USD.

configurations: Option<Vec<Configuration>>

Amazon EMR releases 4.x or later.

The list of configurations supplied for an EMR cluster instance group. You can specify a separate configuration for each instance group (master, core, and task).

ebs_configuration: Option<EbsConfiguration>

EBS configurations that will be attached to each EC2 instance in the instance group.

instance_count: i64

Target number of instances for the instance group.

instance_role: String

The role of the instance group in the cluster.

instance_type: String

The EC2 instance type for all instances in the instance group.

market: Option<String>

Market type of the EC2 instances used to create a cluster node.

name: Option<String>

Friendly name given to the instance group.