Struct rusoto_emr::InstanceGroupDetail

pub struct InstanceGroupDetail {
    pub bid_price: Option<String>,
    pub creation_date_time: f64,
    pub end_date_time: Option<f64>,
    pub instance_group_id: Option<String>,
    pub instance_request_count: i64,
    pub instance_role: String,
    pub instance_running_count: i64,
    pub instance_type: String,
    pub last_state_change_reason: Option<String>,
    pub market: String,
    pub name: Option<String>,
    pub ready_date_time: Option<f64>,
    pub start_date_time: Option<f64>,
    pub state: String,
}

Detailed information about an instance group.

Fields
bid_price: Option<String>

Bid price for EC2 Instances when launching nodes as Spot Instances, expressed in USD.

creation_date_time: f64

The date/time the instance group was created.

end_date_time: Option<f64>

The date/time the instance group was terminated.

instance_group_id: Option<String>

Unique identifier for the instance group.

instance_request_count: i64

Target number of instances to run in the instance group.

instance_role: String

Instance group role in the cluster

instance_running_count: i64

Actual count of running instances.

instance_type: String

EC2 instance type.

last_state_change_reason: Option<String>

Details regarding the state of the instance group.

market: String

Market type of the EC2 instances used to create a cluster node.

name: Option<String>

Friendly name for the instance group.

ready_date_time: Option<f64>

The date/time the instance group was available to the cluster.

start_date_time: Option<f64>

The date/time the instance group was started.

state: String

State of instance group. The following values are deprecated: STARTING, TERMINATED, and FAILED.