Struct rusoto_emr::InstanceTypeConfig

pub struct InstanceTypeConfig {
    pub bid_price: Option<String>,
    pub bid_price_as_percentage_of_on_demand_price: Option<f64>,
    pub configurations: Option<Vec<Configuration>>,
    pub ebs_configuration: Option<EbsConfiguration>,
    pub instance_type: String,
    pub weighted_capacity: Option<i64>,
}

An instance type configuration for each instance type in an instance fleet, which determines the EC2 instances Amazon EMR attempts to provision to fulfill On-Demand and Spot target capacities. There can be a maximum of 5 instance type configurations in a fleet.

The instance fleet configuration is available only in Amazon EMR versions 4.8.0 and later, excluding 5.0.x versions.

Fields
bid_price: Option<String>

The bid price for each EC2 Spot instance type as defined by InstanceType. Expressed in USD. If neither BidPrice nor BidPriceAsPercentageOfOnDemandPrice is provided, BidPriceAsPercentageOfOnDemandPrice defaults to 100%.

bid_price_as_percentage_of_on_demand_price: Option<f64>

The bid price, as a percentage of On-Demand price, for each EC2 Spot instance as defined by InstanceType. Expressed as a number between 0 and 1000 (for example, 20 specifies 20%). If neither BidPrice nor BidPriceAsPercentageOfOnDemandPrice is provided, BidPriceAsPercentageOfOnDemandPrice defaults to 100%.

configurations: Option<Vec<Configuration>>

A configuration classification that applies when provisioning cluster instances, which can include configurations for applications and software that run on the cluster.

ebs_configuration: Option<EbsConfiguration>

The configuration of Amazon Elastic Block Storage (EBS) attached to each instance as defined by InstanceType.

instance_type: String

An EC2 instance type, such as m3.xlarge.

weighted_capacity: Option<i64>

The number of units that a provisioned instance of this type provides toward fulfilling the target capacities defined in InstanceFleetConfig. This value is 1 for a master instance fleet, and must be greater than 0 for core and task instance fleets.