Struct rusoto_emr::JobFlowDetail

pub struct JobFlowDetail {
    pub ami_version: Option<String>,
    pub auto_scaling_role: Option<String>,
    pub bootstrap_actions: Option<Vec<BootstrapActionDetail>>,
    pub execution_status_detail: JobFlowExecutionStatusDetail,
    pub instances: JobFlowInstancesDetail,
    pub job_flow_id: String,
    pub job_flow_role: Option<String>,
    pub log_uri: Option<String>,
    pub name: String,
    pub scale_down_behavior: Option<String>,
    pub service_role: Option<String>,
    pub steps: Option<Vec<StepDetail>>,
    pub supported_products: Option<Vec<String>>,
    pub visible_to_all_users: Option<bool>,
}

A description of a cluster (job flow).

Fields
ami_version: Option<String>

The version of the AMI used to initialize Amazon EC2 instances in the job flow. For a list of AMI versions currently supported by Amazon EMR, see AMI Versions Supported in EMR in the Amazon EMR Developer Guide.

auto_scaling_role: Option<String>

An IAM role for automatic scaling policies. The default role is EMR_AutoScaling_DefaultRole. The IAM role provides a way for the automatic scaling feature to get the required permissions it needs to launch and terminate EC2 instances in an instance group.

bootstrap_actions: Option<Vec<BootstrapActionDetail>>

A list of the bootstrap actions run by the job flow.

execution_status_detail: JobFlowExecutionStatusDetail

Describes the execution status of the job flow.

instances: JobFlowInstancesDetail

Describes the Amazon EC2 instances of the job flow.

job_flow_id: String

The job flow identifier.

job_flow_role: Option<String>

The IAM role that was specified when the job flow was launched. The EC2 instances of the job flow assume this role.

log_uri: Option<String>

The location in Amazon S3 where log files for the job are stored.

name: String

The name of the job flow.

scale_down_behavior: Option<String>

The way that individual Amazon EC2 instances terminate when an automatic scale-in activity occurs or an instance group is resized. TERMINATE_AT_INSTANCE_HOUR indicates that Amazon EMR terminates nodes at the instance-hour boundary, regardless of when the request to terminate the instance was submitted. This option is only available with Amazon EMR 5.1.0 and later and is the default for clusters created using that version. TERMINATE_AT_TASK_COMPLETION indicates that Amazon EMR blacklists and drains tasks from nodes before terminating the Amazon EC2 instances, regardless of the instance-hour boundary. With either behavior, Amazon EMR removes the least active nodes first and blocks instance termination if it could lead to HDFS corruption. TERMINATE_AT_TASK_COMPLETION available only in Amazon EMR version 4.1.0 and later, and is the default for versions of Amazon EMR earlier than 5.1.0.

service_role: Option<String>

The IAM role that will be assumed by the Amazon EMR service to access AWS resources on your behalf.

steps: Option<Vec<StepDetail>>

A list of steps run by the job flow.

supported_products: Option<Vec<String>>

A list of strings set by third party software when the job flow is launched. If you are not using third party software to manage the job flow this value is empty.

visible_to_all_users: Option<bool>

Specifies whether the cluster is visible to all IAM users of the AWS account associated with the cluster. If this value is set to true, all IAM users of that AWS account can view and (if they have the proper policy permissions set) manage the cluster. If it is set to false, only the IAM user that created the cluster can view and manage it. This value can be changed using the SetVisibleToAllUsers action.