Struct rusoto_emr::JobFlowInstancesConfig

pub struct JobFlowInstancesConfig {
    pub additional_master_security_groups: Option<Vec<String>>,
    pub additional_slave_security_groups: Option<Vec<String>>,
    pub ec_2_key_name: Option<String>,
    pub ec_2_subnet_id: Option<String>,
    pub ec_2_subnet_ids: Option<Vec<String>>,
    pub emr_managed_master_security_group: Option<String>,
    pub emr_managed_slave_security_group: Option<String>,
    pub hadoop_version: Option<String>,
    pub instance_count: Option<i64>,
    pub instance_fleets: Option<Vec<InstanceFleetConfig>>,
    pub instance_groups: Option<Vec<InstanceGroupConfig>>,
    pub keep_job_flow_alive_when_no_steps: Option<bool>,
    pub master_instance_type: Option<String>,
    pub placement: Option<PlacementType>,
    pub service_access_security_group: Option<String>,
    pub slave_instance_type: Option<String>,
    pub termination_protected: Option<bool>,
}

A description of the Amazon EC2 instance on which the cluster (job flow) runs. A valid JobFlowInstancesConfig must contain either InstanceGroups or InstanceFleets, which is the recommended configuration. They cannot be used together. You may also have MasterInstanceType, SlaveInstanceType, and InstanceCount (all three must be present), but we don't recommend this configuration.

Fields
additional_master_security_groups: Option<Vec<String>>

A list of additional Amazon EC2 security group IDs for the master node.

additional_slave_security_groups: Option<Vec<String>>

A list of additional Amazon EC2 security group IDs for the slave nodes.

ec_2_key_name: Option<String>

The name of the EC2 key pair that can be used to ssh to the master node as the user called "hadoop."

ec_2_subnet_id: Option<String>

Applies to clusters that use the uniform instance group configuration. To launch the cluster in Amazon Virtual Private Cloud (Amazon VPC), set this parameter to the identifier of the Amazon VPC subnet where you want the cluster to launch. If you do not specify this value, the cluster launches in the normal Amazon Web Services cloud, outside of an Amazon VPC, if the account launching the cluster supports EC2 Classic networks in the region where the cluster launches.

Amazon VPC currently does not support cluster compute quadruple extra large (cc1.4xlarge) instances. Thus you cannot specify the cc1.4xlarge instance type for clusters launched in an Amazon VPC.

ec_2_subnet_ids: Option<Vec<String>>

Applies to clusters that use the instance fleet configuration. When multiple EC2 subnet IDs are specified, Amazon EMR evaluates them and launches instances in the optimal subnet.

The instance fleet configuration is available only in Amazon EMR versions 4.8.0 and later, excluding 5.0.x versions.

emr_managed_master_security_group: Option<String>

The identifier of the Amazon EC2 security group for the master node.

emr_managed_slave_security_group: Option<String>

The identifier of the Amazon EC2 security group for the slave nodes.

hadoop_version: Option<String>

The Hadoop version for the cluster. Valid inputs are "0.18" (deprecated), "0.20" (deprecated), "0.20.205" (deprecated), "1.0.3", "2.2.0", or "2.4.0". If you do not set this value, the default of 0.18 is used, unless the AmiVersion parameter is set in the RunJobFlow call, in which case the default version of Hadoop for that AMI version is used.

instance_count: Option<i64>

The number of EC2 instances in the cluster.

instance_fleets: Option<Vec<InstanceFleetConfig>>

The instance fleet configuration is available only in Amazon EMR versions 4.8.0 and later, excluding 5.0.x versions.

Describes the EC2 instances and instance configurations for clusters that use the instance fleet configuration.

instance_groups: Option<Vec<InstanceGroupConfig>>

Configuration for the instance groups in a cluster.

keep_job_flow_alive_when_no_steps: Option<bool>

Specifies whether the cluster should remain available after completing all steps.

master_instance_type: Option<String>

The EC2 instance type of the master node.

placement: Option<PlacementType>

The Availability Zone in which the cluster runs.

service_access_security_group: Option<String>

The identifier of the Amazon EC2 security group for the Amazon EMR service to access clusters in VPC private subnets.

slave_instance_type: Option<String>

The EC2 instance type of the slave nodes.

termination_protected: Option<bool>

Specifies whether to lock the cluster to prevent the Amazon EC2 instances from being terminated by API call, user intervention, or in the event of a job-flow error.