Struct rusoto_emr::JobFlowInstancesDetail

pub struct JobFlowInstancesDetail {
    pub ec_2_key_name: Option<String>,
    pub ec_2_subnet_id: Option<String>,
    pub hadoop_version: Option<String>,
    pub instance_count: i64,
    pub instance_groups: Option<Vec<InstanceGroupDetail>>,
    pub keep_job_flow_alive_when_no_steps: Option<bool>,
    pub master_instance_id: Option<String>,
    pub master_instance_type: String,
    pub master_public_dns_name: Option<String>,
    pub normalized_instance_hours: Option<i64>,
    pub placement: Option<PlacementType>,
    pub slave_instance_type: String,
    pub termination_protected: Option<bool>,
}

Specify the type of Amazon EC2 instances that the cluster (job flow) runs on.

Fields
ec_2_key_name: Option<String>

The name of an Amazon EC2 key pair that can be used to ssh to the master node.

ec_2_subnet_id: Option<String>

For clusters launched within Amazon Virtual Private Cloud, this is the identifier of the subnet where the cluster was launched.

hadoop_version: Option<String>

The Hadoop version for the cluster.

instance_count: i64

The number of Amazon EC2 instances in the cluster. If the value is 1, the same instance serves as both the master and slave node. If the value is greater than 1, one instance is the master node and all others are slave nodes.

instance_groups: Option<Vec<InstanceGroupDetail>>

Details about the instance groups in a cluster.

keep_job_flow_alive_when_no_steps: Option<bool>

Specifies whether the cluster should remain available after completing all steps.

master_instance_id: Option<String>

The Amazon EC2 instance identifier of the master node.

master_instance_type: String

The Amazon EC2 master node instance type.

master_public_dns_name: Option<String>

The DNS name of the master node.

normalized_instance_hours: Option<i64>

An approximation of the cost of the cluster, represented in m1.small/hours. This value is incremented one time for every hour that an m1.small runs. Larger instances are weighted more, so an Amazon EC2 instance that is roughly four times more expensive would result in the normalized instance hours being incremented by four. This result is only an approximation and does not reflect the actual billing rate.

placement: Option<PlacementType>

The Amazon EC2 Availability Zone for the cluster.

slave_instance_type: String

The Amazon EC2 slave node instance type.

termination_protected: Option<bool>

Specifies whether the Amazon EC2 instances in the cluster are protected from termination by API calls, user intervention, or in the event of a job-flow error.