Struct rusoto_emr::KerberosAttributes

pub struct KerberosAttributes {
    pub ad_domain_join_password: Option<String>,
    pub ad_domain_join_user: Option<String>,
    pub cross_realm_trust_principal_password: Option<String>,
    pub kdc_admin_password: String,
    pub realm: String,
}

Attributes for Kerberos configuration when Kerberos authentication is enabled using a security configuration. For more information see Use Kerberos Authentication in the EMR Management Guide.

Fields
ad_domain_join_password: Option<String>

The Active Directory password for ADDomainJoinUser.

ad_domain_join_user: Option<String>

Required only when establishing a cross-realm trust with an Active Directory domain. A user with sufficient privileges to join resources to the domain.

cross_realm_trust_principal_password: Option<String>

Required only when establishing a cross-realm trust with a KDC in a different realm. The cross-realm principal password, which must be identical across realms.

kdc_admin_password: String

The password used within the cluster for the kadmin service on the cluster-dedicated KDC, which maintains Kerberos principals, password policies, and keytabs for the cluster.

realm: String

The name of the Kerberos realm to which all nodes in a cluster belong. For example, EC2.INTERNAL.