Struct rusoto_emr::Cluster

pub struct Cluster {
    pub applications: Option<Vec<Application>>,
    pub auto_scaling_role: Option<String>,
    pub auto_terminate: Option<bool>,
    pub configurations: Option<Vec<Configuration>>,
    pub custom_ami_id: Option<String>,
    pub ebs_root_volume_size: Option<i64>,
    pub ec_2_instance_attributes: Option<Ec2InstanceAttributes>,
    pub id: Option<String>,
    pub instance_collection_type: Option<String>,
    pub kerberos_attributes: Option<KerberosAttributes>,
    pub log_uri: Option<String>,
    pub master_public_dns_name: Option<String>,
    pub name: Option<String>,
    pub normalized_instance_hours: Option<i64>,
    pub release_label: Option<String>,
    pub repo_upgrade_on_boot: Option<String>,
    pub requested_ami_version: Option<String>,
    pub running_ami_version: Option<String>,
    pub scale_down_behavior: Option<String>,
    pub security_configuration: Option<String>,
    pub service_role: Option<String>,
    pub status: Option<ClusterStatus>,
    pub tags: Option<Vec<Tag>>,
    pub termination_protected: Option<bool>,
    pub visible_to_all_users: Option<bool>,
}

The detailed description of the cluster.

Fields
applications: Option<Vec<Application>>

The applications installed on this cluster.

auto_scaling_role: Option<String>

An IAM role for automatic scaling policies. The default role is EMR_AutoScaling_DefaultRole. The IAM role provides permissions that the automatic scaling feature requires to launch and terminate EC2 instances in an instance group.

auto_terminate: Option<bool>

Specifies whether the cluster should terminate after completing all steps.

configurations: Option<Vec<Configuration>>

Applies only to Amazon EMR releases 4.x and later. The list of Configurations supplied to the EMR cluster.

custom_ami_id: Option<String>

Available only in Amazon EMR version 5.7.0 and later. The ID of a custom Amazon EBS-backed Linux AMI if the cluster uses a custom AMI.

ebs_root_volume_size: Option<i64>

The size, in GiB, of the EBS root device volume of the Linux AMI that is used for each EC2 instance. Available in Amazon EMR version 4.x and later.

ec_2_instance_attributes: Option<Ec2InstanceAttributes>

Provides information about the EC2 instances in a cluster grouped by category. For example, key name, subnet ID, IAM instance profile, and so on.

id: Option<String>

The unique identifier for the cluster.

instance_collection_type: Option<String>

The instance fleet configuration is available only in Amazon EMR versions 4.8.0 and later, excluding 5.0.x versions.

The instance group configuration of the cluster. A value of INSTANCEGROUP indicates a uniform instance group configuration. A value of INSTANCEFLEET indicates an instance fleets configuration.

kerberos_attributes: Option<KerberosAttributes>

Attributes for Kerberos configuration when Kerberos authentication is enabled using a security configuration. For more information see Use Kerberos Authentication in the EMR Management Guide.

log_uri: Option<String>

The path to the Amazon S3 location where logs for this cluster are stored.

master_public_dns_name: Option<String>

The DNS name of the master node. If the cluster is on a private subnet, this is the private DNS name. On a public subnet, this is the public DNS name.

name: Option<String>

The name of the cluster.

normalized_instance_hours: Option<i64>

An approximation of the cost of the cluster, represented in m1.small/hours. This value is incremented one time for every hour an m1.small instance runs. Larger instances are weighted more, so an EC2 instance that is roughly four times more expensive would result in the normalized instance hours being incremented by four. This result is only an approximation and does not reflect the actual billing rate.

release_label: Option<String>

The release label for the Amazon EMR release.

repo_upgrade_on_boot: Option<String>

Applies only when CustomAmiID is used. Specifies the type of updates that are applied from the Amazon Linux AMI package repositories when an instance boots using the AMI.

requested_ami_version: Option<String>

The AMI version requested for this cluster.

running_ami_version: Option<String>

The AMI version running on this cluster.

scale_down_behavior: Option<String>

The way that individual Amazon EC2 instances terminate when an automatic scale-in activity occurs or an instance group is resized. TERMINATE_AT_INSTANCE_HOUR indicates that Amazon EMR terminates nodes at the instance-hour boundary, regardless of when the request to terminate the instance was submitted. This option is only available with Amazon EMR 5.1.0 and later and is the default for clusters created using that version. TERMINATE_AT_TASK_COMPLETION indicates that Amazon EMR blacklists and drains tasks from nodes before terminating the Amazon EC2 instances, regardless of the instance-hour boundary. With either behavior, Amazon EMR removes the least active nodes first and blocks instance termination if it could lead to HDFS corruption. TERMINATE_AT_TASK_COMPLETION is available only in Amazon EMR version 4.1.0 and later, and is the default for versions of Amazon EMR earlier than 5.1.0.

security_configuration: Option<String>

The name of the security configuration applied to the cluster.

service_role: Option<String>

The IAM role that will be assumed by the Amazon EMR service to access AWS resources on your behalf.

status: Option<ClusterStatus>

The current status details about the cluster.

tags: Option<Vec<Tag>>

A list of tags associated with a cluster.

termination_protected: Option<bool>

Indicates whether Amazon EMR will lock the cluster to prevent the EC2 instances from being terminated by an API call or user intervention, or in the event of a cluster error.

visible_to_all_users: Option<bool>

Indicates whether the cluster is visible to all IAM users of the AWS account associated with the cluster. If this value is set to true, all IAM users of that AWS account can view and manage the cluster if they have the proper policy permissions set. If this value is false, only the IAM user that created the cluster can view and manage it. This value can be changed using the SetVisibleToAllUsers action.