Struct rusoto_emr::Ec2InstanceAttributes

pub struct Ec2InstanceAttributes {
    pub additional_master_security_groups: Option<Vec<String>>,
    pub additional_slave_security_groups: Option<Vec<String>>,
    pub ec_2_availability_zone: Option<String>,
    pub ec_2_key_name: Option<String>,
    pub ec_2_subnet_id: Option<String>,
    pub emr_managed_master_security_group: Option<String>,
    pub emr_managed_slave_security_group: Option<String>,
    pub iam_instance_profile: Option<String>,
    pub requested_ec_2_availability_zones: Option<Vec<String>>,
    pub requested_ec_2_subnet_ids: Option<Vec<String>>,
    pub service_access_security_group: Option<String>,
}

Provides information about the EC2 instances in a cluster grouped by category. For example, key name, subnet ID, IAM instance profile, and so on.

Fields
additional_master_security_groups: Option<Vec<String>>

A list of additional Amazon EC2 security group IDs for the master node.

additional_slave_security_groups: Option<Vec<String>>

A list of additional Amazon EC2 security group IDs for the slave nodes.

ec_2_availability_zone: Option<String>

The Availability Zone in which the cluster will run.

ec_2_key_name: Option<String>

The name of the Amazon EC2 key pair to use when connecting with SSH into the master node as a user named "hadoop".

ec_2_subnet_id: Option<String>

To launch the cluster in Amazon VPC, set this parameter to the identifier of the Amazon VPC subnet where you want the cluster to launch. If you do not specify this value, the cluster is launched in the normal AWS cloud, outside of a VPC.

Amazon VPC currently does not support cluster compute quadruple extra large (cc1.4xlarge) instances. Thus, you cannot specify the cc1.4xlarge instance type for nodes of a cluster launched in a VPC.

emr_managed_master_security_group: Option<String>

The identifier of the Amazon EC2 security group for the master node.

emr_managed_slave_security_group: Option<String>

The identifier of the Amazon EC2 security group for the slave nodes.

iam_instance_profile: Option<String>

The IAM role that was specified when the cluster was launched. The EC2 instances of the cluster assume this role.

requested_ec_2_availability_zones: Option<Vec<String>>

Applies to clusters configured with the instance fleets option. Specifies one or more Availability Zones in which to launch EC2 cluster instances when the EC2-Classic network configuration is supported. Amazon EMR chooses the Availability Zone with the best fit from among the list of RequestedEc2AvailabilityZones, and then launches all cluster instances within that Availability Zone. If you do not specify this value, Amazon EMR chooses the Availability Zone for you. RequestedEc2SubnetIDs and RequestedEc2AvailabilityZones cannot be specified together.

requested_ec_2_subnet_ids: Option<Vec<String>>

Applies to clusters configured with the instance fleets option. Specifies the unique identifier of one or more Amazon EC2 subnets in which to launch EC2 cluster instances. Subnets must exist within the same VPC. Amazon EMR chooses the EC2 subnet with the best fit from among the list of RequestedEc2SubnetIds, and then launches all cluster instances within that Subnet. If this value is not specified, and the account and region support EC2-Classic networks, the cluster launches instances in the EC2-Classic network and uses RequestedEc2AvailabilityZones instead of this setting. If EC2-Classic is not supported, and no Subnet is specified, Amazon EMR chooses the subnet for you. RequestedEc2SubnetIDs and RequestedEc2AvailabilityZones cannot be specified together.

service_access_security_group: Option<String>

The identifier of the Amazon EC2 security group for the Amazon EMR service to access clusters in VPC private subnets.