Struct rusoto_emr::EmrClient

pub struct EmrClient { /* fields omitted */ }

A client for the Amazon EMR API.

Methods

impl EmrClient

pub fn new(region: Region) -> EmrClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> EmrClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,