Struct rusoto_emr::RunJobFlowInput

pub struct RunJobFlowInput {
    pub additional_info: Option<String>,
    pub ami_version: Option<String>,
    pub applications: Option<Vec<Application>>,
    pub auto_scaling_role: Option<String>,
    pub bootstrap_actions: Option<Vec<BootstrapActionConfig>>,
    pub configurations: Option<Vec<Configuration>>,
    pub custom_ami_id: Option<String>,
    pub ebs_root_volume_size: Option<i64>,
    pub instances: JobFlowInstancesConfig,
    pub job_flow_role: Option<String>,
    pub kerberos_attributes: Option<KerberosAttributes>,
    pub log_uri: Option<String>,
    pub name: String,
    pub new_supported_products: Option<Vec<SupportedProductConfig>>,
    pub release_label: Option<String>,
    pub repo_upgrade_on_boot: Option<String>,
    pub scale_down_behavior: Option<String>,
    pub security_configuration: Option<String>,
    pub service_role: Option<String>,
    pub steps: Option<Vec<StepConfig>>,
    pub supported_products: Option<Vec<String>>,
    pub tags: Option<Vec<Tag>>,
    pub visible_to_all_users: Option<bool>,
}

Input to the RunJobFlow operation.

Fields
additional_info: Option<String>

A JSON string for selecting additional features.

ami_version: Option<String>

Applies only to Amazon EMR AMI versions 3.x and 2.x. For Amazon EMR releases 4.0 and later, ReleaseLabel is used. To specify a custom AMI, use CustomAmiID.

applications: Option<Vec<Application>>

For Amazon EMR releases 4.0 and later. A list of applications for the cluster. Valid values are: "Hadoop", "Hive", "Mahout", "Pig", and "Spark." They are case insensitive.

auto_scaling_role: Option<String>

An IAM role for automatic scaling policies. The default role is EMR_AutoScaling_DefaultRole. The IAM role provides permissions that the automatic scaling feature requires to launch and terminate EC2 instances in an instance group.

bootstrap_actions: Option<Vec<BootstrapActionConfig>>

A list of bootstrap actions to run before Hadoop starts on the cluster nodes.

configurations: Option<Vec<Configuration>>

For Amazon EMR releases 4.0 and later. The list of configurations supplied for the EMR cluster you are creating.

custom_ami_id: Option<String>

Available only in Amazon EMR version 5.7.0 and later. The ID of a custom Amazon EBS-backed Linux AMI. If specified, Amazon EMR uses this AMI when it launches cluster EC2 instances. For more information about custom AMIs in Amazon EMR, see Using a Custom AMI in the Amazon EMR Management Guide. If omitted, the cluster uses the base Linux AMI for the ReleaseLabel specified. For Amazon EMR versions 2.x and 3.x, use AmiVersion instead.

For information about creating a custom AMI, see Creating an Amazon EBS-Backed Linux AMI in the Amazon Elastic Compute Cloud User Guide for Linux Instances. For information about finding an AMI ID, see Finding a Linux AMI.

ebs_root_volume_size: Option<i64>

The size, in GiB, of the EBS root device volume of the Linux AMI that is used for each EC2 instance. Available in Amazon EMR version 4.x and later.

instances: JobFlowInstancesConfig

A specification of the number and type of Amazon EC2 instances.

job_flow_role: Option<String>

Also called instance profile and EC2 role. An IAM role for an EMR cluster. The EC2 instances of the cluster assume this role. The default role is EMR_EC2_DefaultRole. In order to use the default role, you must have already created it using the CLI or console.

kerberos_attributes: Option<KerberosAttributes>

Attributes for Kerberos configuration when Kerberos authentication is enabled using a security configuration. For more information see Use Kerberos Authentication in the EMR Management Guide.

log_uri: Option<String>

The location in Amazon S3 to write the log files of the job flow. If a value is not provided, logs are not created.

name: String

The name of the job flow.

new_supported_products: Option<Vec<SupportedProductConfig>>

For Amazon EMR releases 3.x and 2.x. For Amazon EMR releases 4.x and later, use Applications.

A list of strings that indicates third-party software to use with the job flow that accepts a user argument list. EMR accepts and forwards the argument list to the corresponding installation script as bootstrap action arguments. For more information, see "Launch a Job Flow on the MapR Distribution for Hadoop" in the Amazon EMR Developer Guide. Supported values are:

"mapr-m3" - launch the cluster using MapR M3 Edition.

"mapr-m5" - launch the cluster using MapR M5 Edition.

"mapr" with the user arguments specifying "--edition,m3" or "--edition,m5" - launch the job flow using MapR M3 or M5 Edition respectively.

"mapr-m7" - launch the cluster using MapR M7 Edition.

"hunk" - launch the cluster with the Hunk Big Data Analtics Platform.

"hue"- launch the cluster with Hue installed.

"spark" - launch the cluster with Apache Spark installed.

"ganglia" - launch the cluster with the Ganglia Monitoring System installed.

release_label: Option<String>

The Amazon EMR release label, which determines the version of open-source application packages installed on the cluster. Release labels are in the form emr-x.x.x, where x.x.x is an Amazon EMR release version, for example, emr-5.14.0. For more information about Amazon EMR release versions and included application versions and features, see http://docs.aws.amazon.com/emr/latest/ReleaseGuide/. The release label applies only to Amazon EMR releases versions 4.x and later. Earlier versions use AmiVersion.

repo_upgrade_on_boot: Option<String>

Applies only when CustomAmiID is used. Specifies which updates from the Amazon Linux AMI package repositories to apply automatically when the instance boots using the AMI. If omitted, the default is SECURITY, which indicates that only security updates are applied. If NONE is specified, no updates are applied, and all updates must be applied manually.

scale_down_behavior: Option<String>

Specifies the way that individual Amazon EC2 instances terminate when an automatic scale-in activity occurs or an instance group is resized. TERMINATE_AT_INSTANCE_HOUR indicates that Amazon EMR terminates nodes at the instance-hour boundary, regardless of when the request to terminate the instance was submitted. This option is only available with Amazon EMR 5.1.0 and later and is the default for clusters created using that version. TERMINATE_AT_TASK_COMPLETION indicates that Amazon EMR blacklists and drains tasks from nodes before terminating the Amazon EC2 instances, regardless of the instance-hour boundary. With either behavior, Amazon EMR removes the least active nodes first and blocks instance termination if it could lead to HDFS corruption. TERMINATE_AT_TASK_COMPLETION available only in Amazon EMR version 4.1.0 and later, and is the default for versions of Amazon EMR earlier than 5.1.0.

security_configuration: Option<String>

The name of a security configuration to apply to the cluster.

service_role: Option<String>

The IAM role that will be assumed by the Amazon EMR service to access AWS resources on your behalf.

steps: Option<Vec<StepConfig>>

A list of steps to run.

supported_products: Option<Vec<String>>

For Amazon EMR releases 3.x and 2.x. For Amazon EMR releases 4.x and later, use Applications.

A list of strings that indicates third-party software to use. For more information, see the Amazon EMR Developer Guide. Currently supported values are:

"mapr-m3" - launch the job flow using MapR M3 Edition.

"mapr-m5" - launch the job flow using MapR M5 Edition.

tags: Option<Vec<Tag>>

A list of tags to associate with a cluster and propagate to Amazon EC2 instances.

visible_to_all_users: Option<bool>

Whether the cluster is visible to all IAM users of the AWS account associated with the cluster. If this value is set to true, all IAM users of that AWS account can view and (if they have the proper policy permissions set) manage the cluster. If it is set to false, only the IAM user that created the cluster can view and manage it.