Trait rusoto_emr::Emr

pub trait Emr {
    fn add_instance_fleet(
        &self, 
        input: AddInstanceFleetInput
    ) -> RusotoFuture<AddInstanceFleetOutput, AddInstanceFleetError>;

    fn add_instance_groups(
        &self, 
        input: AddInstanceGroupsInput
    ) -> RusotoFuture<AddInstanceGroupsOutput, AddInstanceGroupsError>;

    fn add_job_flow_steps(
        &self, 
        input: AddJobFlowStepsInput
    ) -> RusotoFuture<AddJobFlowStepsOutput, AddJobFlowStepsError>;

    fn add_tags(
        &self, 
        input: AddTagsInput
    ) -> RusotoFuture<AddTagsOutput, AddTagsError>;

    fn cancel_steps(
        &self, 
        input: CancelStepsInput
    ) -> RusotoFuture<CancelStepsOutput, CancelStepsError>;

    fn create_security_configuration(
        &self, 
        input: CreateSecurityConfigurationInput
    ) -> RusotoFuture<CreateSecurityConfigurationOutput, CreateSecurityConfigurationError>;

    fn delete_security_configuration(
        &self, 
        input: DeleteSecurityConfigurationInput
    ) -> RusotoFuture<DeleteSecurityConfigurationOutput, DeleteSecurityConfigurationError>;

    fn describe_cluster(
        &self, 
        input: DescribeClusterInput
    ) -> RusotoFuture<DescribeClusterOutput, DescribeClusterError>;

    fn describe_job_flows(
        &self, 
        input: DescribeJobFlowsInput
    ) -> RusotoFuture<DescribeJobFlowsOutput, DescribeJobFlowsError>;

    fn describe_security_configuration(
        &self, 
        input: DescribeSecurityConfigurationInput
    ) -> RusotoFuture<DescribeSecurityConfigurationOutput, DescribeSecurityConfigurationError>;

    fn describe_step(
        &self, 
        input: DescribeStepInput
    ) -> RusotoFuture<DescribeStepOutput, DescribeStepError>;

    fn list_bootstrap_actions(
        &self, 
        input: ListBootstrapActionsInput
    ) -> RusotoFuture<ListBootstrapActionsOutput, ListBootstrapActionsError>;

    fn list_clusters(
        &self, 
        input: ListClustersInput
    ) -> RusotoFuture<ListClustersOutput, ListClustersError>;

    fn list_instance_fleets(
        &self, 
        input: ListInstanceFleetsInput
    ) -> RusotoFuture<ListInstanceFleetsOutput, ListInstanceFleetsError>;

    fn list_instance_groups(
        &self, 
        input: ListInstanceGroupsInput
    ) -> RusotoFuture<ListInstanceGroupsOutput, ListInstanceGroupsError>;

    fn list_instances(
        &self, 
        input: ListInstancesInput
    ) -> RusotoFuture<ListInstancesOutput, ListInstancesError>;

    fn list_security_configurations(
        &self, 
        input: ListSecurityConfigurationsInput
    ) -> RusotoFuture<ListSecurityConfigurationsOutput, ListSecurityConfigurationsError>;

    fn list_steps(
        &self, 
        input: ListStepsInput
    ) -> RusotoFuture<ListStepsOutput, ListStepsError>;

    fn modify_instance_fleet(
        &self, 
        input: ModifyInstanceFleetInput
    ) -> RusotoFuture<(), ModifyInstanceFleetError>;

    fn modify_instance_groups(
        &self, 
        input: ModifyInstanceGroupsInput
    ) -> RusotoFuture<(), ModifyInstanceGroupsError>;

    fn put_auto_scaling_policy(
        &self, 
        input: PutAutoScalingPolicyInput
    ) -> RusotoFuture<PutAutoScalingPolicyOutput, PutAutoScalingPolicyError>;

    fn remove_auto_scaling_policy(
        &self, 
        input: RemoveAutoScalingPolicyInput
    ) -> RusotoFuture<RemoveAutoScalingPolicyOutput, RemoveAutoScalingPolicyError>;

    fn remove_tags(
        &self, 
        input: RemoveTagsInput
    ) -> RusotoFuture<RemoveTagsOutput, RemoveTagsError>;

    fn run_job_flow(
        &self, 
        input: RunJobFlowInput
    ) -> RusotoFuture<RunJobFlowOutput, RunJobFlowError>;

    fn set_termination_protection(
        &self, 
        input: SetTerminationProtectionInput
    ) -> RusotoFuture<(), SetTerminationProtectionError>;

    fn set_visible_to_all_users(
        &self, 
        input: SetVisibleToAllUsersInput
    ) -> RusotoFuture<(), SetVisibleToAllUsersError>;

    fn terminate_job_flows(
        &self, 
        input: TerminateJobFlowsInput
    ) -> RusotoFuture<(), TerminateJobFlowsError>;
}

Trait representing the capabilities of the Amazon EMR API. Amazon EMR clients implement this trait.

Required methods

fn add_instance_fleet(
&self,
input: AddInstanceFleetInput
) -> RusotoFuture<AddInstanceFleetOutput, AddInstanceFleetError>

Adds an instance fleet to a running cluster.

The instance fleet configuration is available only in Amazon EMR versions 4.8.0 and later, excluding 5.0.x.

fn add_instance_groups(
&self,
input: AddInstanceGroupsInput
) -> RusotoFuture<AddInstanceGroupsOutput, AddInstanceGroupsError>

Adds one or more instance groups to a running cluster.

fn add_job_flow_steps(
&self,
input: AddJobFlowStepsInput
) -> RusotoFuture<AddJobFlowStepsOutput, AddJobFlowStepsError>

AddJobFlowSteps adds new steps to a running cluster. A maximum of 256 steps are allowed in each job flow.

If your cluster is long-running (such as a Hive data warehouse) or complex, you may require more than 256 steps to process your data. You can bypass the 256-step limitation in various ways, including using SSH to connect to the master node and submitting queries directly to the software running on the master node, such as Hive and Hadoop. For more information on how to do this, see Add More than 256 Steps to a Cluster in the Amazon EMR Management Guide.

A step specifies the location of a JAR file stored either on the master node of the cluster or in Amazon S3. Each step is performed by the main function of the main class of the JAR file. The main class can be specified either in the manifest of the JAR or by using the MainFunction parameter of the step.

Amazon EMR executes each step in the order listed. For a step to be considered complete, the main function must exit with a zero exit code and all Hadoop jobs started while the step was running must have completed and run successfully.

You can only add steps to a cluster that is in one of the following states: STARTING, BOOTSTRAPPING, RUNNING, or WAITING.

fn add_tags(
&self,
input: AddTagsInput
) -> RusotoFuture<AddTagsOutput, AddTagsError>

Adds tags to an Amazon EMR resource. Tags make it easier to associate clusters in various ways, such as grouping clusters to track your Amazon EMR resource allocation costs. For more information, see Tag Clusters.

fn cancel_steps(
&self,
input: CancelStepsInput
) -> RusotoFuture<CancelStepsOutput, CancelStepsError>

Cancels a pending step or steps in a running cluster. Available only in Amazon EMR versions 4.8.0 and later, excluding version 5.0.0. A maximum of 256 steps are allowed in each CancelSteps request. CancelSteps is idempotent but asynchronous; it does not guarantee a step will be canceled, even if the request is successfully submitted. You can only cancel steps that are in a PENDING state.

fn create_security_configuration(
&self,
input: CreateSecurityConfigurationInput
) -> RusotoFuture<CreateSecurityConfigurationOutput, CreateSecurityConfigurationError>

Creates a security configuration, which is stored in the service and can be specified when a cluster is created.

fn delete_security_configuration(
&self,
input: DeleteSecurityConfigurationInput
) -> RusotoFuture<DeleteSecurityConfigurationOutput, DeleteSecurityConfigurationError>

Deletes a security configuration.

fn describe_cluster(
&self,
input: DescribeClusterInput
) -> RusotoFuture<DescribeClusterOutput, DescribeClusterError>

Provides cluster-level details including status, hardware and software configuration, VPC settings, and so on.

fn describe_job_flows(
&self,
input: DescribeJobFlowsInput
) -> RusotoFuture<DescribeJobFlowsOutput, DescribeJobFlowsError>

This API is deprecated and will eventually be removed. We recommend you use ListClusters, DescribeCluster, ListSteps, ListInstanceGroups and ListBootstrapActions instead.

DescribeJobFlows returns a list of job flows that match all of the supplied parameters. The parameters can include a list of job flow IDs, job flow states, and restrictions on job flow creation date and time.

Regardless of supplied parameters, only job flows created within the last two months are returned.

If no parameters are supplied, then job flows matching either of the following criteria are returned:

Job flows created and completed in the last two weeks

Job flows created within the last two months that are in one of the following states: RUNNING, WAITING, SHUTTING_DOWN, STARTING

Amazon EMR can return a maximum of 512 job flow descriptions.

fn describe_security_configuration(
&self,
input: DescribeSecurityConfigurationInput
) -> RusotoFuture<DescribeSecurityConfigurationOutput, DescribeSecurityConfigurationError>

Provides the details of a security configuration by returning the configuration JSON.

fn describe_step(
&self,
input: DescribeStepInput
) -> RusotoFuture<DescribeStepOutput, DescribeStepError>

Provides more detail about the cluster step.

fn list_bootstrap_actions(
&self,
input: ListBootstrapActionsInput
) -> RusotoFuture<ListBootstrapActionsOutput, ListBootstrapActionsError>

Provides information about the bootstrap actions associated with a cluster.

fn list_clusters(
&self,
input: ListClustersInput
) -> RusotoFuture<ListClustersOutput, ListClustersError>

Provides the status of all clusters visible to this AWS account. Allows you to filter the list of clusters based on certain criteria; for example, filtering by cluster creation date and time or by status. This call returns a maximum of 50 clusters per call, but returns a marker to track the paging of the cluster list across multiple ListClusters calls.

fn list_instance_fleets(
&self,
input: ListInstanceFleetsInput
) -> RusotoFuture<ListInstanceFleetsOutput, ListInstanceFleetsError>

Lists all available details about the instance fleets in a cluster.

The instance fleet configuration is available only in Amazon EMR versions 4.8.0 and later, excluding 5.0.x versions.

fn list_instance_groups(
&self,
input: ListInstanceGroupsInput
) -> RusotoFuture<ListInstanceGroupsOutput, ListInstanceGroupsError>

Provides all available details about the instance groups in a cluster.

fn list_instances(
&self,
input: ListInstancesInput
) -> RusotoFuture<ListInstancesOutput, ListInstancesError>

Provides information for all active EC2 instances and EC2 instances terminated in the last 30 days, up to a maximum of 2,000. EC2 instances in any of the following states are considered active: AWAITING_FULFILLMENT, PROVISIONING, BOOTSTRAPPING, RUNNING.

fn list_security_configurations(
&self,
input: ListSecurityConfigurationsInput
) -> RusotoFuture<ListSecurityConfigurationsOutput, ListSecurityConfigurationsError>

Lists all the security configurations visible to this account, providing their creation dates and times, and their names. This call returns a maximum of 50 clusters per call, but returns a marker to track the paging of the cluster list across multiple ListSecurityConfigurations calls.

fn list_steps(
&self,
input: ListStepsInput
) -> RusotoFuture<ListStepsOutput, ListStepsError>

Provides a list of steps for the cluster in reverse order unless you specify stepIds with the request.

fn modify_instance_fleet(
&self,
input: ModifyInstanceFleetInput
) -> RusotoFuture<(), ModifyInstanceFleetError>

Modifies the target On-Demand and target Spot capacities for the instance fleet with the specified InstanceFleetID within the cluster specified using ClusterID. The call either succeeds or fails atomically.

The instance fleet configuration is available only in Amazon EMR versions 4.8.0 and later, excluding 5.0.x versions.

fn modify_instance_groups(
&self,
input: ModifyInstanceGroupsInput
) -> RusotoFuture<(), ModifyInstanceGroupsError>

ModifyInstanceGroups modifies the number of nodes and configuration settings of an instance group. The input parameters include the new target instance count for the group and the instance group ID. The call will either succeed or fail atomically.

fn put_auto_scaling_policy(
&self,
input: PutAutoScalingPolicyInput
) -> RusotoFuture<PutAutoScalingPolicyOutput, PutAutoScalingPolicyError>

Creates or updates an automatic scaling policy for a core instance group or task instance group in an Amazon EMR cluster. The automatic scaling policy defines how an instance group dynamically adds and terminates EC2 instances in response to the value of a CloudWatch metric.

fn remove_auto_scaling_policy(
&self,
input: RemoveAutoScalingPolicyInput
) -> RusotoFuture<RemoveAutoScalingPolicyOutput, RemoveAutoScalingPolicyError>

Removes an automatic scaling policy from a specified instance group within an EMR cluster.

fn remove_tags(
&self,
input: RemoveTagsInput
) -> RusotoFuture<RemoveTagsOutput, RemoveTagsError>

Removes tags from an Amazon EMR resource. Tags make it easier to associate clusters in various ways, such as grouping clusters to track your Amazon EMR resource allocation costs. For more information, see Tag Clusters.

The following example removes the stack tag with value Prod from a cluster:

fn run_job_flow(
&self,
input: RunJobFlowInput
) -> RusotoFuture<RunJobFlowOutput, RunJobFlowError>

RunJobFlow creates and starts running a new cluster (job flow). The cluster runs the steps specified. After the steps complete, the cluster stops and the HDFS partition is lost. To prevent loss of data, configure the last step of the job flow to store results in Amazon S3. If the JobFlowInstancesConfig KeepJobFlowAliveWhenNoSteps parameter is set to TRUE, the cluster transitions to the WAITING state rather than shutting down after the steps have completed.

For additional protection, you can set the JobFlowInstancesConfig TerminationProtected parameter to TRUE to lock the cluster and prevent it from being terminated by API call, user intervention, or in the event of a job flow error.

A maximum of 256 steps are allowed in each job flow.

If your cluster is long-running (such as a Hive data warehouse) or complex, you may require more than 256 steps to process your data. You can bypass the 256-step limitation in various ways, including using the SSH shell to connect to the master node and submitting queries directly to the software running on the master node, such as Hive and Hadoop. For more information on how to do this, see Add More than 256 Steps to a Cluster in the Amazon EMR Management Guide.

For long running clusters, we recommend that you periodically store your results.

The instance fleets configuration is available only in Amazon EMR versions 4.8.0 and later, excluding 5.0.x versions. The RunJobFlow request can contain InstanceFleets parameters or InstanceGroups parameters, but not both.

fn set_termination_protection(
&self,
input: SetTerminationProtectionInput
) -> RusotoFuture<(), SetTerminationProtectionError>

SetTerminationProtection locks a cluster (job flow) so the EC2 instances in the cluster cannot be terminated by user intervention, an API call, or in the event of a job-flow error. The cluster still terminates upon successful completion of the job flow. Calling SetTerminationProtection on a cluster is similar to calling the Amazon EC2 DisableAPITermination API on all EC2 instances in a cluster.

SetTerminationProtection is used to prevent accidental termination of a cluster and to ensure that in the event of an error, the instances persist so that you can recover any data stored in their ephemeral instance storage.

To terminate a cluster that has been locked by setting SetTerminationProtection to true, you must first unlock the job flow by a subsequent call to SetTerminationProtection in which you set the value to false.

For more information, seeManaging Cluster Termination in the Amazon EMR Management Guide.

fn set_visible_to_all_users(
&self,
input: SetVisibleToAllUsersInput
) -> RusotoFuture<(), SetVisibleToAllUsersError>

Sets whether all AWS Identity and Access Management (IAM) users under your account can access the specified clusters (job flows). This action works on running clusters. You can also set the visibility of a cluster when you launch it using the VisibleToAllUsers parameter of RunJobFlow. The SetVisibleToAllUsers action can be called only by an IAM user who created the cluster or the AWS account that owns the cluster.

fn terminate_job_flows(
&self,
input: TerminateJobFlowsInput
) -> RusotoFuture<(), TerminateJobFlowsError>

TerminateJobFlows shuts a list of clusters (job flows) down. When a job flow is shut down, any step not yet completed is canceled and the EC2 instances on which the cluster is running are stopped. Any log files not already saved are uploaded to Amazon S3 if a LogUri was specified when the cluster was created.

The maximum number of clusters allowed is 10. The call to TerminateJobFlows is asynchronous. Depending on the configuration of the cluster, it may take up to 1-5 minutes for the cluster to completely terminate and release allocated resources, such as Amazon EC2 instances.

Loading content...